Update Lesson 1: Intro to HTML
HTML stands for Hypertext Markup Language.
The first lesson introduces us to HTML tags. It tells us that the tags start and end with angular brackets(<>) anything inside the angular bracket is the tag name. For example <h1> is the heading tag. The maximum range is from h1 to h6 where 1 indicates the biggest heading and 6 indicates the smallest heading. Every text that is to be entered requires to be between an opening tag and a closing tag. <h1> Hi </h1>
The tag <br> is used for line spacing. <br> can be used within the heading tag like:
<h1> Haan Bhai, <br>
What's up? <h1/>